driver inverse fortran kind

+++ sample/FORTRAN/driver_inverse.f90
@@ -37,10 +37,10 @@
       real(mytype),dimension(:),allocatable:: cosx,cosy,cosz
       real(mytype) pi,twopi,sinyz,diff,cdiff,ccdiff,ans
 
-      integer(8) Ntot
+      integer(i8) Ntot
       real(mytype) factor
-      real(8) rtime1,rtime2,Nglob,prec
-      real(8) gt(12,3),gtcomm(3),tc
+      real(r8) rtime1,rtime2,Nglob,prec
+      real(r8) gt(12,3),gtcomm(3),tc
       integer ierr,nu,ndim,dims(2),nproc,proc_id
       integer istart(3),iend(3),isize(3)
       integer fstart(3),fend(3),fsize(3)
@@ -353,7 +353,7 @@
 
       use p3dfft
 
-      integer(8) nar,i
+      integer(i8) nar,i
       complex(mytype) X(nar)
       real(mytype) f
 
@@ -374,8 +374,8 @@
       use p3dfft
 
       integer x,y,z,proc_id
-      integer(8) i,Nar
-      real(8) Nglob
+      integer(i8) i,Nar
+      real(r8) Nglob
       real(mytype), target :: Ar(1,1,*)
       integer Fstart(3),Fend(3),Fsize(3)
 
@@ -398,10 +398,10 @@
       use p3dfft
 
       integer x,y,z,proc_id
-      integer(8) i,Nar
+      integer(i8) i,Nar
       complex(mytype) Ar(1,1,*)
       integer Fstart(3),Fend(3),Fsize(3)
-      real(8) Nglob
+      real(r8) Nglob
 
       call p3dfft_get_dims(Fstart,Fend,Fsize,2)
